Initial commit for plugin that allows users to upload images
The imagare (image + share) plugin installs a new menu `Tools` ->
`Image Upload` where users can upload images. After the upload the
user gets an URL under which the uploaded image is available. This is
useful for sharing screenshots and linking them from review comments.

There are services in the web that provide the possibility of sharing
screenshots but if working in a corporate environment they cannot be
used since screenshots must not leave the company. This is why it
makes sense to have this functionality as a Gerrit plugin.

With this initial commit it is only possible to upload images by
pasting them from the clipboard into the WebUI. Other possibilities to
upload images (upload from local filesystem, upload by drag and drop)
are not implemented yet.

The uploaded images are stored in the `refs/images/*` namespace of the
`All-Projects` project. From the coding it is already possible to
decide in which project an uploaded image should be stored. E.g. if
later an image upload could be triggered directly from the change
screen we could store the image in the corresponding project.

The plugin should also support to set a size limit for the uploaded
images and provide a way to configure the default project into which
images are stored, but this is not part of this initial commit.

Change-Id: I31a051938dad83444520855c7869e1810140c016
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>