Support decoration of image links in Gerrit WebUI
If links to screenshots are posted in review comments a reviewer needs
to open that link in order to see the image. Instead allow to decorate
the image link so that the image can be seen immediately without
opening the link. Two ways of link decoration are supported:

1. Tooltip:
On mouse over a tooltip with the image is shown.

2. Inline:
The image is inlined and shown instead of the URL.

Change-Id: I4b4e420d2d850aacf5898defb05df52c2ebfdcbb
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>