Allow users to stage images before upload
Currently images are uploaded immediately when selected, pasted or
dropped. Some users may prefer a 2 step process to upload images so
that they can check that the correct pictures are selected before
upload. The image upload has then to be triggered explicitely. Whether
images are staged or directly uploaded can be controlled by
configuration and a user preference.

Change-Id: I52845e12d3d71befc077132bea3162b15173d7ce
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>