Add global capability that allows the deletion of own uploaded images
After the upload of an image a user may relalize that he did a mistake
and uploaded a wrong image. In this case the user wants to delete the
uploaded image, but administrators don't want to allow users to delete
arbitary uploaded images. This is way it makes sense to have a global
capability which allows the deletion of own uploaded images without
need to have force push on the images namespace which would allow the
user to delete any uploaded image.

Change-Id: I5c5bdff02c85070d04cf06269ddebe6d6422805f
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>