Fix retrieval of path from URL
The ImageServlet needs the endcoded URL path, but it used
HttpServletRequest.getPathInfo() to retrieve it which returns the
decoded path. So far this worked because there was a bug in Guice [1]
that made HttpServletRequest.getPathInfo() return the encoded path.
Now this bug was fixed in Guice and Gerrit was updated to use the new
Guice version. Hence the ImageServlet started to fail and needs to be
adapted.

[1] https://github.com/google/guice/issues/745

Change-Id: Ie44786acaa70b9612572727dad856d3fdcf472ae
Signed-off-by: Edwin Kempin <ekempin@google.com>