Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Does this and the parent change need to be tied into the same topic as adding the PG UI?  Or can they be decoupled and submitted already?

This change in principle would not have to be (although I would redo the formatting then based on 'master'). I think however that it would be nice to merge the removal of GWT together with the addition of the new UI to not lose the functionality. Although this is the master-branch this code is used with Gerrit-2.16, which still is able to use GWT. We should however create a dedicated stable-2.16 branch. I don't have the permissions. Can you do that?

Patch-set: 1
Reviewer: Gerrit User 1065256 <1065256@173816e5-2b9a-37c3-8a2e-48639d4f1153>