Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Shouldn't just this change been done on gerrit 3.0+?
> 
> Since 2.16 still supports GWTUI.

I've tried to keep it, when I was starting to work on implementing the new UI, but was unable to make it work, since when using PolyGerrit and trying to open the upload screen, both the GWT-based and the Polymer-based UI were trying to load, thereby interfering.
Note, that the images will still be shown in the comments in the old UI, only uploading them, and changing the user settings won't be available in the old UI anymore.
Since this plugin was pretty much unmaintained and I was assuming only very few people would use the old UI in 2.16, I chose the easy way out to save time.
Do you know a way to have both at the same time? Then I am happy to revert this change on stable-2.16 and make it work, if you want that.

Patch-set: 2